(maint) benchmark: add protocol override; choose host port carefully
Allow selection of the protocol via a --protocol argument that allows
either http or https, and rework the protocol, host, and port defaults
accordingly.  This also fixes the defaults to be dependent on the
protocol.  Previously we might intermingle http and https hosts and
ports.

Change the default host from 127.0.0.1 to localhost so that it'll
match the localhost certificates created by pdbbox.