(PE-36947) Allow for Puppet::Util::Puppetdb to flush config
This provides a hook for flushing the Puppetdb classes @config variable.
This in turn means the next time Puppetdb.config is callled it will
reload the configuration file.
This becomes important if in some command using this utility, we have
accessed the Puppetdb.config, then go on to update puppet's
puppetdb.conf with new information, and need to then take additional
operations with Puppetdb.config that should be based on the new
configuration state.