(PE-36939) query.monitor: report invariant failures more carefully
Adjust each assertion after a bit more consideration, and under the
assumption that we're going to remove with-monitored-execution from
the monitor thread, and that the outright failure of the monitor
shouldn't crash pdb, but just leave it running in a degraded (query
monitorless) state.

Ensure that whatever happens, the failures end up in the log.