(maint) Remove github macos tests
These have been causing intermittent, and not so intermittent failures
that appear to be due to extremely poor performance from the gihub
runners, corroroborated by some online reports which also noted that
they're expensive (per minute?).

Given that and the fact that no one's actively developing puppetdb
from macos right now, remove the tests with the expectation that
if/when we restore them, we'll restore them elsewhere.