(PE-36939) query.monitor: don't restart pdb for monitor failures
If the monitor thread fails, make sure to report it as an error in the
log, but don't allow it to restart pdb.  This does assume that even on
failures, the monitor doesn't entangle other code in ways that could
break queries (or anything else critical).

The current expectation is that for this particular subsystem, since
it is fairly detached, it's better to allow pdb to continue in a
degraded (query monitorless) state until the problem can be fixed with
a new release, than to risk catastrophic reboot loops.