(maint) Preserve random-hosts generated for initial-hosts-channel
When benchmark starts, it generates a list of initial host-maps based on
random selections from the catalog/factset/report sample data. This
means that every time benchmark is run a new ordering of catalogs
and facts are going to be pushed to PuppetDB. This can cause a great
deal of initial load replacing the catalogs and facts wholesale until
PuppetDB has caught up to processing one entire simulated node interval.

This makes it impractical to stop and restart benchmark when working
with large simulated runs.

Benchmark continues to massage these host-maps for each simulation loop,
and makes use of TempFileBuffer for communication of host-maps between
the read and write channels. By default this is written to a generated
tmp directory, and the transitional file state only exists for as long
as it takes the channels to pull them.

By setting a new --simulation-dir, benchmark will instead write every
host-map to that given directory without removing any files.

At startup, when initializing host, any host-# present as a preserved
host-map file will be read from the --simulation-dir rather than
initialized randomly. Any host-# not found as a preserved host-map file
will be initialized randomly as usual.

The final host-map files have been renamed to be just host-# to allow
for this indexing.

Preserved state is loaded in the start-simulation-loop's
async/pipeline-blocking's transformation function to avoid retaining a
reference to the data in the random-hosts sequence.

We then fill out preserved maps where allowed by --facts, --reports,
--catalogs in case they weren't present during previous run(s).

And we remove content from preserved maps to match just the requested
--facts, --reports, --catalogs or --archive content. (The preserved state
may include data not requested by the current run.)