(maint) Sync initial-hosts on benchmark re-execution
When benchmark starts, it generates a list of initial host-maps based on
random selections from the catalog/factset/report sample data. This
means that every time benchmark is run a new set of catalogs and facts
are going to be pushed to PuppetDB. This can cause a great deal of
initial load replacing the catalogs and facts wholesale until PuppetDB
has caught up to processing one entire simulated node interval.

This makes it impractical to stop and restart benchmark when working
with large simulated runs.

This patch attempts to improve this be encoding the certname of the
original sample file in the catalog version string. We then query
catalogs certname, version fields, and form an index of certname ->
original-certname. This index is then used to ensure that the set of
initial hostmaps is regenerated from the same base catalog and facteset
so there should not be excess churn when benchmark begins pushing
commands.