(PE-37113) benchmark: freeze/thaw explicitly, not inside channel
Drop the custom TempFileBuffer backed channel that handles (deferred)
freezing and thawing of the host state internally/implicitly in favor
of a normal channel, and have the simulators freeze and thaw their own
data, explicitly.

Aside from simplifying things, this also stops using core async
operations, and blocking operations (e.g. the poll/take), inside a
core async channel.

Worth fixing regardless, we noticed this because it was causing
performance problems with the attempt to add command-related queries
that were alleviated by increasing the core.async.pool-size (which
shouldn't have been necessary).  The issue was also reported as a
problem by clojure.core.async.go-checking=true.

In the new arrangement, rearrange the channels to distinguish between
the much larger host-info instances that are only needed for the
currently active hosts (those with an active command sequence), and
the list of all hosts that the simulators cycle through.  For the
latter, just track the host-path to the host's frozen data.

So now, at startup, the populate-hosts function returns a lazy
sequence of host-paths for the requested host range, after making sure
that the frozen data exists for each host, either because we're
re-using a --simulation-dir that had the relevant HOST file, or
because populate hosts created a new one for the host.

Those host paths flow into the mixer that the simulators read from,
and then the simulators thaw the data, advance and freeze the new host
state, put the host path on the sim-next channel so it'll come around
again for the next iteration, and put the new host-info on the
host-info-ch for the senders.

Previously, we had the simulators attempt to update the thawed host
state to reflect the current benchmark run (i.e. whether
facts/catalogs/reports were requested as compared to a previous run
that generated the --simulation-dir).  Now we just let the simulation
dir host maps retain all their data and have the simulators prune the
maps to reflect the current run's options before putting them on the
host-info-ch for the senders.

Remove the initial host population testing for now; we'll restore it
later.