(PDB-5713) Add test for negated regex
Add a test case to illustrate the bug. Based on the catalog
wire format, parameters cannot be nil and must at least be the empty
map, so this was a bug in the test data.

Adjusting parameters to be non-nil is important because otherwise the
jsonb existence operator ?  returns null rather than false, which when
negated is still null so it does not result in any matched rows even
when the parentheses are added to the SQL raw.