(PE-37113) benchmark: prevent simultaneous cmds for same host
Previously, when in --nummsgs mode (no rate limit), there was nothing
to prevent a host from "wrapping around" and being processed again,
while its previous command sequence or even *sequences* were still
pending.

Fix that by pulling the simulators out of their own little world and
into a "global" loop with the rest of the processing, i.e. instead of
having the simulators process the host-infos in an isolated circle as
fast as the senders ask for them, close the loop so that the senders
won't see a given host again until the current command sequence has
finished.

After this change a host's information travels from the simulator, to
the sender, and then, only after the final event for that host has
executed, back to the simulators for the next round.