(maint) Ensure key matches when adding random resource to catalog
Previously, would get key from a random resource, then associate that
key with a new random-resource. Since the resources are transformed back to a
list of just the values, the transiently mismatched key isn't a huge
issue, but over a large number of benchmark update-host iterations, it
can lead to a case where you have two resources with the same type/name
pair.

Ex:
 {"{:type A, :name B}" {:type A, :name B, ...}
  "{:type B, :name C}" {:type A, :name B, ...}}

Since the key (B, C) of the added resource doesn't match the resource
(A, B) it is added with, a duplicate type/name pair can make it
through to the updated catalog resources...

And then future update-host calls fail in the :post assertion test
of puppetlabs.puppetdb.catalogs/transform-resources because the
duplicate resource is lost when resources are transformed to a map.