(PE-37113) benchmark: restrict number of concurrently active hosts
Add a --concurrent-hosts option that restricts the number of hosts
that have active command sequences.  Without this, in --nummsgs mode,
there's no bound on the heap size that might be required (as more and
more delayed sequences pile up in the scheduler), and while we might
change this later, right now each active sequence will have the entire
host-info (e.g. possibly factset, catalog, report) in RAM, so there's
a notable limit (perhaps on the order of hundreds of concurrent
sequences per GB).

The maximum number of hosts that could be in flight in --runinterval
mode depends on the interval and host count.  (If --concurrent-hosts
is set too low, then benchmark won't be able to hit the requested
rate.)