(PE-37466) query.monitor: forget queries immediately
Change forget to immediately drop the queries, i.e. don't just
tombstone the entry via :forget and wait for the monitor thread to
clean up.

Although we'd originally tried to reduce the potential for
concurrency-related issues by having the monitor loop handle most of
the work, :forget was an unnecessary complication (after reviewing all
the related operations), and the current implementation was also
somewhat wrong, i.e. next-expired-query should have been checking for
:forget (and immediately dropping the query) before looking at the
deadline.  As a result, it might return a deadline for a query that
was already forgotten as the next wait, delaying the cleanup.  Not
necessarily immediately harmful, but undesirable.