(PE-37466) Ensure query monitor forgets for all relevant routes
Previously, non-streaming queries like those that set the `ast_only`
query param would have query monitoring established via
wrap-typical-query, but then the monitor would never be told to forget
them (either on success or failure).

This could cause the monitor to throw an exception when a subsequent
query came in over the same channel as a query that should have
already been forgotten.  That exception was intentionally thrown
because the monitor data structures are keyed in a way that requires
all active queries to have unique channels.

The streaming query pathway (the most common pathway) has to handle
"forgetting" indirectly in the streaming threads since it returns the
http 200 response after the first row, thereby returning from all the
wrappers, and so we can't unconditionaly forget "on the way out" in
wrap-typical-query.  It might also be difficult to know "on the way
in" whether or not any given query/route is going to be streaming or
"immediate" (e.g. when `ast_only` is set, it's immediate, and when
not, it isn't), so adding a wrapper param, or two wrapper variants
might not be desirable.

As an alternative, just have the query engine streaming code mark the
response with a :puppetlabs.puppetdb.query/streaming key that the
wrapper can use to decide whether or not it should handle asking the
monitor to forget the query.