(PE-37466) query.monitor/disconnected?: report unexpected bytes
Before the previous fix to make sure we forget queries on all relevant
routes, the disconnection check would race with jetty and if the check
won when jetty was also reusing the connection/channel, the
next disconnection check would read the first 4k of the next incoming
query, making a big mess.

Although that bug's been fixed, mention any unexpected bytes in the
log so that next time it'll be more obvious when the query monitor is
acting out of turn.