(PE-37113) benchmark: populate hosts more efficiently
Austin and Josh's testing showed a notable decrease in the maximum
command send rate during startup, and Austin tracked it down to the
unnecessary extra write of the initial host state that this commit
removes.

Instead, just adjust the senders to handle a host info map too, so
that populate-hosts can just pass along the host-map when it already
has it.  Previously it would read it, possibly adjust it (to match the
current benchmark arguments), then write it to a file that would
immediately be re-read.