(PE-37537) Add --rand-catalogs, --rand-facts to benchmark
These options will replace --rand-perc and provide for a flexible way to
specify the average frequency of different degrees of catalog and
factset mutations.

Factsets are difficult to work with as a CHANGE-COUNT because it's a
nested map that would take time to count and randomly select from. It's
easier for now to use the existing code that supplies a percentage
chance to change a leaf as it walks the factsets structure.