(maint) testutils: query indexes by name not column
The docstring says it returns the indexes sorted by table then name of
the index, but in actuality it returned the indexes sorted by table and
column names. A table can have more than one index on the same
column(s), so when this was converted to map it would lose any columns
whose table + column set were the same.