Move deactivated/expired status from certnames table
The certnames table has a lot of contention and churn. In a standard
system every new report writes a new latest_report_id and timestamp
row. Additionally package and catalog_inputs information can optionally
be stored, in which case this table would get written by facts and
catalog storage as well. This level of writes causes a high rate of
table bloat and dirtied cache pages in the shared buffers for the
certnames table.

When a query causes a Seq Scan on the certnames table, such as the PE
Console's status page to get a count of active nodes, the performance
can be highly variable based on the number of dirty cache pages. When no
new cache pages are written, the query can take only a few milliseconds,
but in a highly bloated table with a significant number of dirty pages
the same query was seen to take as long as 15 seconds.

see https://wiki.postgresql.org/wiki/Hint_Bits
and https://www.cybertec-postgresql.com/en/speeding-up-things-with-hint-bits/

Split the deactivated and expired timestamps out into a separate table
in order to provide a table that is much less frequently written to that
has all the data necessary to count the number of active nodes. This
table is only written to when a node is first created, when it is
deactivated, or when it expires. So in the steady-state it has nearly no
writes and none of its buffer pages will be dirtied.