Attempt to activate node only when necessary
This insert ran for every single command that PuppetDB processes to
ensure that the certname exists. Most of the time it will conflict on
the certname and perform a conditional update. In Postgres, an update
acquires a row lock and in order to acquire the row lock, Postgres needs
to write the hint bits and xmax on this row, which dirties the buffer
cache page.

more info on row locks can be found
https://www.cybertec-postgresql.com/en/row-locks-in-postgresql/

The statement should modify data _very_ infrequently. When a certname
does not exist it will insert it. The only other time this operation
does anything is if an agent continues to submit commands after it has
been deactivated or expired. So we opt to query for the data and perform
the check in PuppetDB in order to avoid causing the row lock to be
acquired and the buffer page dirtied.

Because we normally perform this check at repeatable read isolation
level, and we are not using a SELECT FOR UPDATE, this does introduce a
small window for a read-then-update issue. To avoid re-activating a node
that should remain deactivated, continue to use the insert on conflict
do update formulation. In the case where a node is deactivated during
the RTU, but should be reactivated, we rely on the fact that if the node
is truly "active" it will submit another command which will then
reactivate the node.