(PDB-5739) facts-query: change stable||volatile to union
Since we've observed that the || operation can be expensive, and we
don't actually want a full factset map here (we want the individual
top-level subtrees), rewrite it as a union.  This shouldn't change
anything material since pdb ensures the two key sets are disjoint.

This also assumes that it's not important to be compatible with the
existing stable||volatile index here.  The existing query could only
have been using the index if postgres is clever enough to see through
the key (via jsonb_each/.*) operation to identify cases where the
index might be relevant.