(maint) pdb-core-routes: return a GET for /, not a bare handler fn
Return a compojure GET route for "/" rather than a context wrapping a
plain clojure function.  Before this change the "/" context handler
was being selected for all routes, so all routes returned the
dashboard redirection.  This was likely caused by the recent
clj-parent upgrade that pulled compojure to a much newer version.

cf.
  https://github.com/weavejester/compojure/wiki/Nesting-routes
  6416f22a26f5f238033cd9f93bcbd55e47ca6833