(maint) benchmark_test: add additional buffer to rand tests
Set the test numbers to accommodate the most extreme results of 10,000
iterations.

Rather than simply loosen the bounds, we track the number of extra
additions and add a buffer to account for the fact that deletions can
actually be modifications if we only have one resource remaining.

We should be getting a balanced number of additions and deletions by
using rand-nth, but if there's only 1 resource remaining deletes become
a modification. We start with 4 resources, so if we get a string of 7
deletes followed by a string of 7 additions, we have a balanced number
of operations, but we will end with 8 resources. This shifts our
expected end number of resources to the right of a simple binomial
distribution.