(PDB-5745) Don't log stack traces for invalid queries
Previously, when responding to an invalid query (bad syntax, etc.),
where the exception message is the only relevant bit, pdb would log
the full stack trace along with the error message.  Stop that.

Instead of throwing a generic IllegalArgument exception for these
errors, throw a more appropriate/specific ex-info exception.  Catch
that in the relevant places and don't include the stack trace when
logging.  Introduce the concept of a :puppetlabs.puppetdb/known-error?
ex-info kind for this purpose.

Add a new bad-query-ex macro to create the exception.  (It's a macro
so that it captures the correct starting line, which is still reported
by clojure test.)