(maint) acceptance tests: manage rhel pgdg repo manually
We must remain on the older postgresql modules for compatibility with
Puppet 6, but those versions still configure the old pgdg signing keys.
On RHEL, do not manage the pgdg repo via the module, and instead set it
up manually so that the new signing keys are not overwritten by the
postgresql module.