(PDB-5747) handle-resource-insert-sql-ex: support batching
In order to support upcoming changes to batch inserts, adjust the
index error handler to also accept a collection of resources and
report any that might be large enough to have provoked the error.

Even though the actual pg limit is (currently) 8191 bytes, we can't be
certain of the culprit because if nothing else, that limit applies
after possible in-line compression:
https://www.postgresql.org/message-id/8326.1289618101%40sss.pgh.pa.us

Treat a resource as suspicious if the UTF-8 encoded length of any
indexed keys exceeds 8000.  Just use UTF-8 for now since the other
encodings seem unlikely:
https://www.postgresql.org/docs/current/multibyte.html#MULTIBYTE-CHARSET-SUPPORTED