(PDB-5792) Prevent duplicate catalogs
Add a migration to make the catalogs certname index unique.

In order to prevent the constraint addition failing for users who do
have duplicate catalogs, purge any duplicate catalogs. The latest
catalog will be kept. If catalogs have the same producer_timestamp, the
id column is used as a tiebreaker.